pyproject.toml
[project]
name = "tiger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fast-bss-eval",
    "huggingface-hub",
    "librosa",
    "numpy",
    "pandas",
    "pandas-stubs",
    "ptflops",
    "pytorch-lightning",
    "pyyaml",
    "rich",
    "safetensors",
    "scipy-stubs",
    "speechbrain",
    "torch-complex",
    "torch-mir-eval",
    "torch-optimizer",
    "torchcodec",
    "typeguard",
    "wandb",
]

[project.optional-dependencies]
cpu = [
    "torch<=2.8",
    "torchvision",
    "torchaudio",
]
cu118 = [
    "torch<=2.8",
    "torchvision",
    "torchaudio",
]
cu128 = [
    "torch<=2.8",
    "torchvision",
    "torchaudio",
]
rocm63 = [
    "torch<=2.8",
    "torchvision",
    "torchaudio",
    "pytorch-triton-rocm",
]
rocm64 = [
    "torch<=2.8",
    "torchvision",
    "torchaudio",
    "pytorch-triton-rocm",
]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu118" },
        { extra = "cu128" },
        { extra = "rocm63" },
        { extra = "rocm64" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-rocm63", extra = "rocm63" },
    { index = "pytorch-rocm64", extra = "rocm64" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-rocm63", extra = "rocm63" },
    { index = "pytorch-rocm64", extra = "rocm64" },
]
pytorch-triton-rocm = [
    { index = "pytorch-rocm63", extra = "rocm63" },
    { index = "pytorch-rocm64", extra = "rocm64" },
]

torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-rocm63", extra = "rocm63" },
    { index = "pytorch-rocm64", extra = "rocm64" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm63"
url = "https://download.pytorch.org/whl/rocm6.3"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm64"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true
